PR feedback - use isApplePayEnabled check in client, move ApplePay config extensions into module
